requirements.txt
fastapi
uvicorn
python-multipart
pillow
opencv-python
ultralytics
opencv-python-headless
supabase
python-dotenv
numpy
onnxruntime